+++ client/src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./components/Home/Home";
 import { LandingPage } from "./pages/Landing/LandingPage";
-import CreateRecipe from "./components/CreateRecipe/CreateRecipe";
+import { CreateRecipePage } from "./pages/CreateRecipe/CreateRecipe";
 import { DetailsPage } from "./pages/Details/DetailsPage";
 import "./App.css";
 
@@ -13,7 +13,7 @@ function App() {
           <Route exact path="/" element={<LandingPage />} />
           <Route exact path="/home" element={<Home />} />
           <Route exact path="/home/:id" element={<DetailsPage />} />
-          <Route exact path="/create-recipe" element={<CreateRecipe />} />
+          <Route exact path="/create-recipe" element={<CreateRecipePage />} />
         </Routes>
       </div>
     </BrowserRouter>
